+++ app/src/main/java/ac/mdiq/podcini/util/event/FeedTagsChangedEvent.kt
@@ -1,0 +1,3 @@
+package ac.mdiq.podcini.util.event
+
+class FeedTagsChangedEvent